requirements.txt
# Core MCP dependencies
mcp==1.0.0
pydantic==2.5.0
httpx==0.25.2
websockets==12.0

# Data processing and analysis
pandas==2.3.0
numpy==2.1.2
scipy==1.11.4
yfinance==0.2.48
alpha-vantage==3.0.0

# Visualization
plotly>=5.0.0
matplotlib==3.9.2
seaborn==0.13.2

# Machine Learning and AI
scikit-learn==1.3.2
statsmodels==0.14.1

# Deep learning (if needed)
torch==2.1.1
transformers==4.36.0

# File handling
openpyxl==3.1.2
xlsxwriter==3.1.9

# Async support
asyncio
aiofiles==23.2.1

# Configuration
python-dotenv==1.0.1
pyyaml==6.0.1
jinja2==3.1.2

# Logging
loguru==0.7.2

# API integration
requests==2.32.3
fastapi==0.104.1
uvicorn==0.24.0
aiohttp==3.9.1

# Date and time handling
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development tools
black==23.11.0
flake8==6.1.0

# Database (optional for data storage)
sqlite3  # Built-in with Python

# AI Integration
openai==1.3.0

# Technical analysis
ta==0.10.2
TA-Lib==0.4.28

# Streamlit
streamlit>=1.45.0 